% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_data.R
\docType{data}
\name{acs_data}
\alias{acs_data}
\title{County-level demographic data from the American Community Survey}
\description{
Data processed from ACS public use samples available at IPUMS.
Data are county-year-level averages. ACS data cover counties with
100,000+ residents. For counties with fewer residents, county-level
characteristics are estimated by copying the demographic characteristics
of the nearest California county that appears in the ACS.
}
\references{
\href{https://usa.ipums.org/usa/}{IPUMS Data}
}
\keyword{data}
