% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_to_county.R
\docType{data}
\name{city_to_county}
\alias{city_to_county}
\title{California city-to-county crosswalk}
\description{
Used to map WARN tables, which report city of layoff, to county
level for aggregation and estimation. The city-to-county mapping
is scraped from https://www.counties.org/cities-within-each-county.
}
\references{
\href{https://www.counties.org/cities-within-each-county}{counties.org Data}
}
\keyword{data}
