% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\docType{data}
\name{pop}
\alias{pop}
\title{County-level population data}
\description{
Contains processed county-year-level population
estimates downloaded from the Census.
}
\references{
\href{https://www2.census.gov/programs-surveys/popest/datasets/2010-2018/counties/asrh/cc-est2018-alldata-06.csv}{BLS Census Data}
}
\keyword{data}
