% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnBar.R
\name{warnBar}
\alias{warnBar}
\title{Generate bar plots from warnExtract data frame}
\usage{
warnBar(data, by = c("rollup", "locality", "reason"))
}
\arguments{
\item{data}{a data frame of WARn data from \link[jobsec]{warnExtract}.}

\item{by}{a string specifying bar fill categories.}
}
\description{
generates bar plots from data frames that were produced by \link[jobsec]{warnExtract}.
}
\examples{
#extract warn data
df<- warnExtract(start_date = "2018-01-01", end_date = "2019-01-01")
#bar plots
warnBar(df, by = "reason")
}
