% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnDownload.R
\name{warnDownload}
\alias{warnDownload}
\title{Calls warnReader function, scrapes WARN notice data}
\usage{
warnDownload(year = NULL)
}
\arguments{
\item{year}{is an integer vector containing the fiscal years for which the function
will pull WARN reports. The default is to pull the current year report. The
fiscal year begins on July 1 of the specified year and ends on June 30 of the
following year. Fiscal years are integers ranging from 2014 to 2019.}
}
\description{
warnDownload converts a WARN table in PDF format to a data frame.
}
\examples{
   # Download WARN files for current year
   warn_current <- warnDownload(2019)
   
   # Replicate the warnSample dataset provided in the package
   warnSample <- warnDownload(2014:2018)
   
}
