% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnExtract.R
\name{warnExtract}
\alias{warnExtract}
\title{Data extractor for summarizing WARN data.}
\usage{
warnExtract(
  warn_data = warnSample,
  start_date = -Inf,
  end_date = Inf,
  type_date = c("effective_date", "notice_date", "received_date"),
  rollup = c("month", "year"),
  layoff_type = c("Closure Permanent", "Closure Temporary", "Closure Unknown",
    "Layoff Permanent", "Layoff Temporary", "Layoff Unknown"),
  counties = NA
)
}
\arguments{
\item{warn_data}{a WARN data set. uses historical data as default. can also provide data set from \link[jobsec]{warnDownload} or\link[jobsec]{warnModeler}..}

\item{start_date}{a start date.}

\item{end_date}{an end date.}

\item{type_date}{a string specifying layoff reason.}

\item{rollup}{a string specifying what time frame to roll up totals: "month" or "year".}

\item{layoff_type}{a string specifying what layoff types to filter for.}

\item{counties}{a character vector of counties to filter for.}
}
\value{
a data frame of summarized WARN data based on user filtered inputs.
}
\description{
Creates a filtered data set from WARN data to use in plotting functions or as a data table.
}
\examples{
df <- warnExtract(counties = c("Alameda County", 
  "Los Angeles County"), rollup = "year")
df <- warnExtract(start_date = "2018-01-01", end_date = "2018-08-01")
}
