% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnMap.R
\name{warnMap}
\alias{warnMap}
\title{Generate map plots rom warnExtract data frame}
\usage{
warnMap(data)
}
\arguments{
\item{data}{a data frame of WARn data from \link[jobsec]{warnExtract}.}
}
\description{
generates map plots from data frames that were produced by \link[jobsec]{warnExtract}.
}
\examples{
#extract warn data
df<- warnExtract(start_date = "2018-01-01", end_date = "2019-01-01")
#plot maps
warnMap(df)
}
