% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnModeler.R
\name{warnModeler}
\alias{warnModeler}
\title{Estimation using macro data and WARN notices}
\usage{
warnModeler(df = warnSample)
}
\arguments{
\item{df}{is a WARN extract data frame. df must be in the same format
as data downloaded using the warnDownload function and must contain at
least 100 unique county-year observations for reasonable inference}
}
\description{
Merges WARN data to lagged macro variables,
then estimates county level layoffs using lagged macro characteristics.
Returns output from OLS regression and regression tree
}
\examples{
   # Replicate warnPrediction dataset provided in package
   data(warnSample)
   warnPrediction <- warnModeler(warnSample)

}
