% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnNamer.R
\name{warnNamer}
\alias{warnNamer}
\title{Clean WARN data}
\usage{
warnNamer(d)
}
\arguments{
\item{d}{is a list object containing scraped WARN pdf}
}
\description{
CA WARN data tables are published in inconsistent format. The
warnNamer function identifies the format of the WARN table and
names the columns appropriately.
}
