% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnPrediction.R
\docType{data}
\name{warnPrediction}
\alias{warnPrediction}
\title{Predicted WARN data}
\description{
Predictions for 2019-20 layoffs by county, output from
warnModeler (using the warnSample data and seed = 25).
}
\keyword{data}
