% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnReader.R
\name{warnReader}
\alias{warnReader}
\title{Scrape WARN notice data}
\usage{
warnReader(url)
}
\arguments{
\item{url}{is the location of a WARN PDF table}
}
\description{
warnReader converts a WARN table in PDF format to a data frame.
Accepts as argument url = the url location of the target WARN table.
}
