% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnSample.R
\docType{data}
\name{warnSample}
\alias{warnSample}
\title{Sample WARN table extracts}
\description{
A sample of a processed WARN report table covering fiscal years
2014-2018. Identical to the output of warnDownload(2014:2018).
}
\references{
\href{https://www.edd.ca.gov/Jobs_and_Training/Layoff_Services_WARN.htm}{California EDD Data}
}
\keyword{data}
